+++ sdgl/backends/sdl3/Sdl3Backend.h
@@ -14,7 +14,7 @@ namespace sdgl::backends::sdl3 {
         bool init() override;
 
         [[nodiscard]]
-        double getAppTime() const override;
+        float getAppTime() const override;
 
         [[nodiscard]]
         string_view name() const override { return "SDL3"; };
